encap: Consolidate MPIVectorEncaps into VectorEncaps.
